Removed remaining `int16' and replace them by EIF_TYPE_INDEX.
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@71796 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/C/run-time/eif_plug.h
@@ -106,7 +106,7 @@ extern long *esize;		/* Size of object given DType */
 /* Conformance query in class GENERAL */
 #define econfg(obj1, obj2) \
 	(((EIF_REFERENCE) obj1 == (EIF_REFERENCE) 0) || ((EIF_REFERENCE) obj2 == (EIF_REFERENCE) 0 )) ? EIF_FALSE: \
-		eif_gen_conf((int16) Dftype(obj2), (int16) Dftype(obj1))
+		eif_gen_conf(Dftype(obj2), Dftype(obj1))
   
 /* Are dynamic types of `obj1' and `obj2' identical? */
 #define estypeg(obj1, obj2) \
